b6

+++ minecraft/net/minecraft/src/FontRenderer.java
@@ -310,7 +310,7 @@ public class FontRenderer implements ResourceManagerReloadListener
      * Draws the specified string.
      */
     
-    public void drawCenteredString(FontRenderer var1, String var2, int var3, int var4, int var5) {
+    public void drawCenteredString(FontRenderer var1, String var2, float var3, float var4, int var5) {
     	var1.drawStringWithShadow(var2, var3 - var1.getStringWidth(var2) / 2, var4, var5);
     }
